pyproject.toml
[project]
name = "ocr-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langfuse>=3.9.3",
    "pillow>=12.0.0",
    "pydantic-ai-slim[logfire,openai]>=1.15.0",
    "pymupdf>=1.26.6",
    "pymupdf-layout>=1.26.6",
    "pymupdf4llm>=0.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
